+++ Changelog.md
@@ -26,6 +26,7 @@
 * Remove unused `Webdriver#get_element_number_by_text`
 * Remove unused `Webdriver#click_on_locator_by_action`
 * Remove unused `Webdriver#select_text_from_page`
+* Remove unused `Webdriver#move_to_one_of_several_displayed_element`
 
 ## 0.1.2 (2020-02-18)